Faces are now properly accounted for in vtkCmbMesh and partially tested.
The test will call meshing of the model faces.  The test will
pass for now as I ignore the fact that the faces aren't getting
meshed.  When Rob implements model face/polygon meshing this
will get changed to check for that but I didn't want to have it
fail when I know it should fail for this and not catch something
else that introduces a bug that would make this fail.