First real step in client-server code for SimBuilderMeshing.
There still needs to be work done for the xml and cmake files
but the client side and server side code should be in good shape.
The operators will still need some more work but I'm hoping it
will be easier when I get together with Yumin to integrate
this work into the gui.