Fixed model entity meshing logic and actual meshed values.
The logic is now consistent that if a model entity has
a 0 value set for a mesh parameter or its boundary isn't
meshed, it will still return true for BuildModelEntityMesh()
and BuildMesh() since the value returned true that it was
able to properly do what it was told even though it was
told not to build a mesh (or delete a mesh if it existed).
Updated the documentation to be clearer as well.