Extract out conceptual model library as a standalone lib.
This is the first pass on seperating out ERDC conceptual model library
into a stand alone discrete model library (vtkDiscreteModel), and a
vtkCMBModel library. The vtkDiscreteModel library will eventually be a
separate open-source(BSD) repo, that can be used outside of ERDC-CMB repo, and be combined
with mesh-server (SiMBA) repo as part of a framework for remote
model-meshing service (Another project sponsored by DOE). The vtkCMBModel
library will still be part of ERDC-CMB repo, which has CMB-specific codes
that are not open-sourced, for now.

The next step is to rename all CMB-classes in vtkDiscreteModel.