Changing way that SMTK gets grid information from CMB.
Instead of using a callback method I am now using an abstract class
defined in SMTK that Model keeps a shared pointer to. The abstract
class has pure virtual functions to get the information and the
implementation of this class is done in CMB and passed to the
smtk Model object.