A bunch of clean up for SMTK.
Fixes:
1) Tooltip for input widgets to show min, max and/or default value
2) Tooltip for BriefDescription for widget label.
3) Option to have Advanced labels as bold and/or italic (not
yet used by the GUI though since it's not being set by the
template file yet).
4) Fixed DefaultIndex being parsed properly from the template file.
5) Double Min/Max with non inclusive property now uses the next
bit as the constraint instead of a default tolerance. For example,
if 0 was a min and the range wasn't inclusive the GUI would
allow the smallest non-zero value represented by a double to
be used for the Qt validator.
6) VoidItem now gets Advanced font from qtUIManager.