Fix issue where widget allows only a single association.
For domain set associations with SimBuilder materials or bc
group associations with SimBuilder boundary conditions, the
constraint was wrong in that it would only allow a single
ModelBuilder object to be associated with a SimBuilder object.
This changes allows multiple domain sets to be associated with
a SimBuilder material or multiple bc goup assocations with a
SimBuilder boundary condition in the GUI widget.