Proper initialization when running inside valgrind.
Setting the environment variable `SMTK_IN_VALGRIND` when
running a program linked to SMTK will cause the library
to initialize the UUID generator differently in order to
prevent **many** defects from being reported. However,
the initialization potentially reduces the quality of
random numbers used for UUIDs, so only define this for
testing.