Fixing attribute uniqueness issue.
If an attribute definition was not set to be unique, it wouldn't write that
out to the instance file and when it was read in by default an
attribute definition was specified to be unique. Thus writing and then reading
an sbi file would change all non-unique attribute definitions to
be unique.