Round out the cursor class rogue's gallery.
This adds dimension-specific subclasses of
UseEntity and ShellEntity to round out the
list of cursor classes. (There are still a
few more we could add if we need to handle
space-time elements, but we don't support
those yet.)

Things compile and tests pass but there is
a lot of work left to get these classes
working properly with real solid models.