Attribute associations for new and old-style models.
Provide methods and storage to link attributes to new-style model entities.
The attribute manager should always be provided a pointer to an instance of
`smtk::model::Storage` so that associations/disassociations can provide
the model an opportunity to respond to attribute changes.
However, it should be possible to associate and disassociate attributes
to/from model entities (UUIDs) without the pointer being set.

Model Storage instances now track (but do not own)
a pointer to an attribute manager. It is NULL by default.
The return value of attachAttribute() methods has changed
so tests are modified to reflect that.