First "simple" model presentation.
The SimpleModelSubphrases class now passes over cell-use
records when presenting model information. It only displays
cells, not uses or shells. In the future, we can add the
shells of the first use as an optional display so that the
structure is visible.

Also, in the future, we can switch to the more complex
view that includes use-records by exposing the `m_passOverUses`
ivar of the generator.