Progress on a VTK-to-CGM-FacetedSurface converter.
Create a `vtk-to-cholla` program that reads a VTK XML
PolyData (`.vtp`) file and writes a
CGM/Facet/Cholla (`.facet` or `.cub`) file which can
then be run through `cgm-convert` to generate an SMTK JSON model.
The program accepts model-face array names. If no per-cell
model-face array is present, each triangle/quad in the polydata
is treated as a separate model face.

The `cgm-convert` application has been modified to initialize
CGMA with a different engine if the `FACET_TYPE` filetype is
requested. This is required as the OCC engine cannot read facet
files.