Prettify subphrases; sort them.
This does destroy some information that can be useful,
such as the ordering of edges within a face. But it
makes for much easier-to-digest presentations of group
members.

Also, we sort by descriptive phrase type, entity type,
entity dimension, and finally title (with special
attention paid to numeric differences) to preserve
groupings.