Use the new SMTK test-data repository.
Add a CMake cache entry `SMTK_DATA_DIR` that can be set to
a clone of the SMTK test data repository at:

    git clone git@www.kitware.com:SMTKTestData.git

and use test data from that directory. Add a ReadMe for the
SMTK project.

Remove the `cube.json` and `cube-commented.json` data from
the model testing directory as it was outdated and should
live in the test-data repository if/when it is updated.

This also fixes a problem with ImportJSON where arrangements
created by default for an entity were being appended with
arrangements from the JSON-data instead of overwritten by
those from the JSON-data.