Removed UIManager singleton pattern.
The singleton pattern of UIManager was removed so that applications can
create as many UIManagers as they want. The singleton pattern was
inherited from SimBuilder3.0, but since now the uiManager is part of the
smtk library, the pattern isn't desired anymore. However, applications can
still use singleton pattern if they wrapp the uiManager with another
class.