Progress on removing arrangements and entities.
Provide `unarrangeEntity()` and `erase()` methods on `Storage`.
These methods are minimally tested; while they remove an
arrangement, that removal will likely leave the solid in
an inconsistent state. For example, removing a vertex that
is part of a chain forming an edge will leave that chain with
a single vertex.

Care must be taken to keep the model in a consistent state.
However, it is unclear that cascading changes is the best
approach since removing low-dimensional components like
vertices would result in the removal of entire shells or
connected components.