First pass at Operator and Parameter classes.
The `smtk::model::Operator` class is a abstract base class
for implementing solid modeling operations.
This pass implements the minimum required to expose parameters
to operators and report on the success or failure of an operation.
The only functionality beyond minimum requirements is that observers
may be registered for parameter changes and will/did-operate events.