Added slot to listen to more signals from QLineEdit
The UI is setup to listen to QLineEdit'seditingFinished signal so that it does not update the related attribute while user is typing. The attribute will be updated after editing is done. However, this will be a problem if the setText() method is used programatically, which is the case when we run ctest. Now we added slot to listen to textChanged() signal which will be fired when setText() is called.