First pass of embedded editor for attribute ref widget.
The idea here is to allow user edit in place the attribute items directly for those referenced attributes. This is useful and convenient when an application only uses instanced view, not the attribute view, which is where normally attributes are edited. We still need to sync the multiple widgets who are referencing the same attribute when the attribute is changed from one widget.