Made label lenth a bit smarter for instanced view.
The attribute item labels vary quite a bit, and a fixed global label length does not always look so well, and wastes a lot of horizontal space if the labels are shorter than the global length, though they do line up nicely. Some logic is added so that we go through the attribute items first to find the longest label as a hint combined with the global length to decide the appropriate length for labels.