Fix + improve RST2HTML user-guide/tutorial targets.
This unbreaks the `smtk_add_doc()` function which would
attempt to create the same target each time it was called.
It also adds the ability to list files included in the main
RST file and have them included as dependencies. Finally,
it will accept a list of figure files that should be copied
into the build directory when modified (so that the HTML
can be viewed properly without installation) and adds install
rules for the HTML and any figures you list.

Some figures and a skeleton for a user guide are included.