Fixed a problem with updating table widget in qtGroupItem
For extensible group items, if the group items has a child group item,
creating new subgroups in this scenario is not showing the whole widget in
the table cell. I added a signal to the qtItem so that when a child group
item's widget size is changed, the parent table widget will force repaint.