Changed the layout of cmb bridge and discrete model dirs.
All the cmb-model and discrete-model files are now under
smtk-bridge-cmb, to be the same place as the cmb-bridge classes,
and the smtkcmb dir is removed. As the first pass,
the cmb-bridge classes and cmb-model classes are built into one lib,
just as before in cmb repo. Eventually, we could separate the cmb-model
files out as a standalone model kernel, and cmb-bridge will only has
bridge-related classes.