Introducing CellTypes, CellTraits, DimensionTypes, and QueryTypes to mesh.
Rather than only have concrete cell classes, we now also have enums. The enums
are easier to handle when you need to query the system as the caller. But
the CellTypes make it easier to write optimized code.

I expect we are going to iterate over QueryTypes a couple of times before
we land on exactly what we want.