Added option to build model bridges as paraview plugins.
Added a cmake option to build model bridges (currently cgm and cmb) as paraview plugins. This means smtk will reqire Paraview to build if this option is turned on. Since we also have a cmake option to build smtk with only VTK,the toplevel cmake logic is separated into two routes for Paraview and VTK separately. Be careful not to intermixthe two routes, otherwise cmake will complain about exported targets or modules not found. These brige plugins are different from traditional paraview plugin in a sense that all they do is to invoke the model kernel on server so that they can register with model manager.