Added try-catch around setting cgm default engines.
There is an exception being throw when calling cgm methods to set its default engine, which in turn causing the Qt application to crash due to unhandled exception in QEventLoop. The default engine seems still being set though. A try-catch is added around the code block so that the application will not crash.