Third batch of changes for remote bridges via Remus.
The smtk-remote-model program now accepts many command-line options,
including some that provide a way to generate a Remus Worker (.RW)
file describing a bridge supported by the program (which bridges are
available depend on the configuration options for your build of SMTK).

These Remus Worker files are now the only way that Remus remote
bridges are registered with the BridgeRegistrar. However, this
is more flexible than the static advertisements that previous
commits were focused on providing and can include site-specific
information.

Eventually the installation step of SMTK will generate RemusWorker
files for supported bridges as templates.

The smtk-model-server program now accepts command-line arguments
as well, including directories to search for Remus Worker files.

The code compiles but the unitRemusBridgeConnection test may hang
trying to connect to a worker that the built-in Remus server
does not start (due to mismatched work name, tag, and requirements
type and/or info).