Fifth batch of changes for remote bridges via Remus.
Get rid of RemusModelBridgeType; it is no longer needed because
Remus does not require a class for input and output mesh types.

Since mesh types are now strings and pre-constructor
options can be set via each bridge class's staticSetup(),
there is no need for this class any more.
Instead, we register the Remus-worker-file-specified mesh type,
creating a duplicate bridge constructor for the modeling kernel
specified in the Remus worker file under the name specified in
the file. This makes it possible for applications to distinguish
local, in-process modeling kernels from Remus kernels.

If there's a requirements file, use it.
Otherwise, the worker won't match the client's description.

Debug registration of bridges.