CGM transcription and some CGM operators.
The CGM bridge now provides basic transcription of a model.
It does nothing fancy (no loops, shells, uses, etc.).
It is not incremental (the entire model is transcribed at once).
But it is enough to get a few operators going.

A few new operators are provided for testing purposes including
the "union" boolean, which expects ModelEntity instances to be
associated to it.

Now smtk::model::Operator provides some syntactic sugar to make
life easier: you can ask for parameters and fiddle with associations
on the underlying attribute without having to ask for the specification.
Also, the Operator version of "associatedEntitiesAs" actually works
because it temporarily sets the attribute System's model manager.

There are still some issues.

+ The input bodies are not removed except by "force majeure" inside
  the implementation of the operation. This is going to be problematic
  in the future as many operators will need to do this manually when
  really it should be automatic (but how can body deletion be detected
  in CGM?).
+ The tessellation of the output body appears to include a bunch of junk
  interior to the body. The outer surface looks dandy, though.
+ Really, Attribute::associatedEntitiesAs<T>() should be fixed by
  setting a model Manager up on the attribute's System. But it is nice
  to keep thinking of the Bridge as possibly working with multiple
  model Manager instances.