Bridge-filetypes to include model engine info.
The 'bridge-filetypes' used to only return a list of file types, and now it return a map
from the model 'engine' to the list of file extensions. The reason behind this is that sometimes
application needs the model engine info to give user hints for selecting bridge and engine to
load certain files because a file type could be supported from different bridge-engines.