Make bridge sessions full-fledged entities.
Bridge sessions now have a cursor class and an entity type
in `smtk/model/EntityTypeBits.h`. Registering a bridge with
a model manager adds an Entity record to the Manager so that
properties such as the name can be retrieved.
Furthermore, this means that bridges can be part of a
hierarchy of DescriptivePhrases via the EntityPhrase class.

Note that this does **not** do away with the Bridge class
itself, which is still required and used to manage operators
and communication across processes.
However, it does make access to bridges simpler, since now
they are tied to a single Manager instance which owns them.
Bridges in turn own ModelEntity instances.