Get Remus bridge minimally usable in Python. Kinda.
This gets a few more methods of RemusBridgeConnection wrapped
so that a process-local Remus server can be started with
application-provided search-directories (where Remus-worker
files will be found).

This also changes the way the remote bridge is advertised
in python; instead of importing it separately from the main
smtk module, it is now imported when the main module is imported.
If the remote bridge cannot be imported, this fails silently
and still allows access to the main smtk module. If you
see `dir(smtk.bridge)` includes `remote`, then you should
be able to:

    import smtk
    rbc = smtk.bridge.remote.RemusBridgeConnection.create()
    rbc.addSearchDir('/usr/local/var/smtk/workers')
    rbc.connectToServer('local', 50505)
    print rbc.bridgeNames()
    print smtk.model.Manager.bridgeNames()
    mgr = smtk.model.Manager.create()
    sess = mgr.createSession('smtk::model[cgm{OpenCascade}]')

There are still several problems:

+ Worker types are not enumerated immediately upon connection
  to a server. Instead, RemusBridgeConnection::bridgeNames()
  must be called.
+ When the BridgeRegistrar **is** notified of new workers,
  it is not done correctly: only a single worker type is
  added and it is named `smtk::model[native]@tcp://127.0.0.1:50505`
  instead of both `smtk::model[cgm{OpenCascade}]` and
  `smtk::model[discrete]` being registered.
+ Calling a model::Manager's createSession method with one of
  the valid bridge names appears to create a zombie process
  and the Python interpreter hangs.