Adding support to expose model operators from tree view.
If user right-clicks on the 'bridge-session' item in the tree view, a context menu will popup with all operator names supported by the associated bridge, and when user clicks on the operator in the context menu, a dialog with UI for the operator specification will show up. Once user clicks OK after filling out the info on the dialog, a signal will be emitted with the specified operator, and the application can connect to the signal, then run the operation as desired. In the future, the context menu will be a little smarter in that based on what type of entity is selected, the menu will show different operations.