Consequences of Bridges being owned by Manager.
Now that Bridge instances are officially owned by a
model Manager instance, their op() method does not
require a reference to the manager... the Bridge can
provide it.

However, this also causes issues because the default
bridge cannot be created in the constructor of BRepModel --
since it is the base class for the model Manager, it has
not yet been promoted to a Manager and thus cannot hand out
shared pointers to its descendant. But this shared pointer
is required by the default Bridge so it can reference its
parent. So, we move construction so that it is on demand
(whenever a model without a bridge asks for a bridge).
This is preferable since we might not wish to advertise a
"default" bridge session in the GUI.

Finally, we add the equality method to Cursor subclasses
so that comparison of cursors works... for some reason it
stopped working (or was working brokenly) prior to this.
Perhaps adding the hash method to the class changed how
shiboken generated the Python class template for each wrapper?