Fixes, better attention to detail for member masks.
Some fixes and enhancements for the code processing
membership mask specifications.

This also improves the Exodus bridge to include the
true embedding dimension of the model (obtained from the
reader), both directly and in the dimension bits of
boundary and domain groups. This in turn causes
more precise descriptions of the groups and gets
the CMBv4 hydra template working so that only groups
of the proper dimensionality appear as available.

Finally, it improves the method by which the
qtAssociationWidget obtains its list of available
entities. Iterating over subphrases is not an efficient
(nor guaranteed accurate) method for obtaining all of
the entities available.