Add "entities" property to result attribute def.
Whenever possible, operators should use the "entities"
item to return model entities affected by the operation.
At some point in the future, any entities listed in this
"entities" item will be marked as dangling by the bridge
before the results are passed on to the caller so that
transcription can occur. (Some bridges like forwarding
bridges may "pre-fetch" the transcription and encode
in JSON to avoid the overhead of many round-trips asking
for each entity to be transcribed.)