Finalize use of ModelEntityItem to store associations.
Previous changes stored the association membership mask
in a ModelEntityItemDefinition. This commit stores the
actual array of associated entities in an item derived
from that definition.

This also consolidates membership testing between
attribute associations and model-entity attribute items.

Finally some membership tests are fixed.
For example, groups should be allowed to associate to
an attribute missing the "group" bit from its membership mask
(as long as it is also missing the "no group" bit).