Support universal operators; add one (set property).
This commit adds support for "universal" operators that can work on any bridge.
It is accomplished by allowing operators to be registered with the base Bridge
class and adding the inheritsOperators() method to bridges so that remote
bridges can avoid inheriting any operators from base classes.

This commit also adds a new operator that is (suprise!) universal:
the "set property" operator.
It sets property values such as the name and color on model entities, which
is typically information not stored in the foreign modeling kernel.