Add "flush records to stream" option to Logger.
Now, in addition to storing log records, each one is
flushed to an ostream you specify as soon as it is
recorded. If you request it, the logger will take
ownership of the stream.

This is useful for debugging purposes where crashes
may prevent later inspection of the log.

Finally, this fixes an issue with smtkInfoMacro
(it was marking the severity as DEBUG, not INFO).