Make Cursors hold weak refs to their model Manager.
Since it was possible for an invalid Cursor to have a null/invalid
shared pointer to the model manager, all the checks for this being
a weak reference already existed.

Although Cursors owning strong references to the Manager was not
an error, it did not philosophically match the spirit of the class.
This change also exposed some separate memory management issues
with the discrete and remote bridges -- which this commit fixes.