Fixed a bug in Session destructor.
The session has a map of UUIDs to vtkModelItem, and has observers on
these items. In the session destrutor, it tries to remove the observers,
and it will crash if any of those vtkModelItem(s) is invalid. Two changes
are added to fix this. One is to use vtkWeakPointer in the map; second is
to listen to the DeleteEvent of vtkModelItem, and remove it from the map.