Merge model's BRepModel and Manager classes.
There is still a need for some cleanup of methods that take
UUIDs instead of EntityRef (cursor) instances.

The main reason for the commit is that often the BRepModel
implementations had to cast to Manager internally in order
to access arrangements, trigger callbacks, etc., which
complicated managing entities.

Doing this now so that cleanup of session registration can
use simpler EntityRef-based API.