Add "assign names" property to model Operator.
When set (it is optional and its default value is OFF),
entity names will be assigned after each operation is
performed.

Names will only be assigned to models listed explicitly
by UUID in the operator result's "entities" item.

This is an efficient way for user interfaces to control
"prettification" of names.