Add convenience method for vector-valued items.
This adds a templated `setValues(I begin, I end)` method to
ModelEntityItem and ValueItemTemplate.
It can be used with vector iterators but not set iterators
because `end - begin` must be defined and return an integer
so that the smtk::attribute::Item's numberOfItems can be set.

This commit also uses the above to fix the CGM boolean operators
so that they properly list the expunged entities.