A new "simple" python API and many fixes.
This commit adds an "smtk.simple" Python module that
provides a basic, function-oriented API for solid modeling
using SMTK's operators. Some of the tests now use this module.

The CGM operator now provides a convenience method for
converting all of the model entities in an item into CGM
references of a given type. The boolean operators now use
this method.

The ModelEntityItem class now provides an `isExtensible()`
method as a convenience for accessing the definition's extensibility.
This is used to fix the XML attribute parser, which was not properly
resizing ModelEntityItem instances that are extensible.

This commit adds a new Sweep operator to the CGM bridge which
allows construction by extrusion, revolution, or sweeps along paths.
It is only partially tested by `cgmBuildUp.py`.