Expose CGM's sweep helix op and fix bugs.
This branch includes initial support for helical sweeps,
but be aware that only Cubit/ACIS support this kind of
sweep; the CGM OpenCascade backend does not yet perform
sweeps.

This also fixes a bug in `ModelEntityItem::appendValue()`.
This bug causes issues when `Operator::associateEntity()` is
called and the association has a minimum number of expected
values; because there are not default UUID values, the first
numberOfValues() entries are null UUIDs; the associated entity
is appended to the end of this array, which leaves null values
at the front of the array. So, even though it is slow, we must
verify that existing slots for model entities are actually in use.

Finally, this branch tweaks the smtk.simple API a little bit:

* Add CreateCylinder to simple python API.
* Solid modeling operations print the error
  log when an operation does not succeed.