Add CreateBody operator and clean up CGM ops.
Adding CreateBody prepares us for boolean operations on
sheet bodies (so that we can do polygonal modeling).

The rest of the cleanup moves a fair amount of repeated
code out of the individual operators to the CGM operator
superclass.

Finally, this restores names and URLs to CGM models by
setting the properties in the read operator.