Added more features to qtModelEntityItem for entity association.
Added a button to automatically associate the ModelEntityItem to selected
entities if they match the memebership mask. Added another button to clear
all entity associations of the item. Another feature is the items in the
popup of the combobox will be hightlighted when mouse over them, and a signal
will be emitted while the highlighted item changes, which in turn will trigger
selection of the corresponding entity in the render view and tree view. This
way the user will know visually which entity is being associated.