Link model tree view text edting with SetOperator
When user edits text in the model tree view to change entity names,
the UI changes need to be linked with SetOperator to change entity names
on the server. Otherwise the changes will be lost once entites are
refeched from the server after certain operations. The same logic applies
to color and visibility changes.