Replace result "entities" w/ "created", "modified".
Operator results now break reported entities into

+ created,
+ modified, and
+ expunged.

Not all of the discrete operators have been changed to reflect this,
but the read operator has so that tests pass.

This also adds methods to the Operator base class:

    void addEntityToResult(OperatorResult, const EntityRef&, ResultEntityOrigin);

    template<typename T>
    void addEntitiesToResult(OperatorResult, const T&, ResultEntityOrigin);

which add entities to the OperatorResult's "modified" or "created"
items based on the ResultEntityOrigin argument.