Disable CreateEdgesOp in discrete.
The transcribing from vtkDiscreteModel to smtk model has some problems after the
vtkCreateEdgesOperator is done and new edges are created. Most likely this is due to
the topolgy relationship of the new edges is not what the smtk trancriber expected, so
the new edges was not transcribed properly. I am disabling it for now until the bug is fixed.