Add APIs to automatically associate operator model entity item.
When a model operator's UI is launched, we want the ModelEntityItem(s) in
the operator specifiction to be automatically associated with the current selection
in the modelView, if the entity flags match with the item's entity mask. If we did not
find any suitable matches in the children of the selected entities, we will also search
the parents. For example, if a Face is selected, but the ModelEnityItem is requiring a model,
we will search up the tree and associated the item with the model owning the face.