Refactor EntityGroupOp in discrete to also handle volome.
In ModelBuilder-v3, there is a vtkModelMaterial class and its operator class,
vtkMaterialOperator. Essentially this is a partitioned group, meaning entities
can only belong to one group (in V3, the group is called Domain set, which is
normally associated with material type of attributes). In this commit, we brought
back the vtkMaterialOperator, so that users can create group of volumes (3d) or group
of faces (2d), then associate (material) attribute to them. However, ideally we
should have a univeral group op across all sessions to cluster entities, and then
these groups can be used for attribute associations.