Add geometry style (parametric/discrete) to models.
This commit adds a new enum named `ModelGeometryStyle` used to
indicate whether a model is discrete or parametric (or some other
type in the future — maybe point clouds, implicit models like RBFs
or other non-solid-models). The various sessions mark their models
with the appropriate type. It is stored as an integer property.

This also adds some new image tests to the various sessions.

Finally, it adds SetEntityProperty to the simple Python API
and uses that in the image tests to set colors on groups.