Fix displayModel test; antialias when interactive.
The displayModel test started failing when we started using
newer VTK versions because the balloon label owned by the view
was turned on by default and rendered as a translucent rectangle
at the lower left corner of the screen.

Turn on antialiasing (multisampling) when entering interactive mode.
This will not affect batch tests and makes the display programs
look much nicer when used interactively.