Added recentValues list in FileItem.
The idea here is that the FileItem will keep a list of recent values that has
been set to the item, so that the UI widget can display the list to the user for convenience.
For example, after a user selected different files for the item a few times, these files will
be displayed in the dropdown list, and can be reselected without going through the file dialog again.