Refactor APIs for association between Entity and Attribute.
Before this change, model manager had one and only one reference to an attribute system, so when we do entity assoication with attribute from different attribute system, such as model operators or simulation templates, we have to switch back and forth the attribute system of the model manager.
This added extra burden for application to manage the logic for switching the system, also could cause hard-to-find bugs if the application logic is not properly handled. So we removed the reference to attribute system from model manager, and changed the association api to include the attributesystem. The manager stills contains the map of entity and its associations for easy access from smtk, but now the association api eliminate the ambiguilty of which attribute system to use.