Removed code for adding Group field array in model source.
The code for adding Group field array is moved to another filter in cmb, vtkSMTKModelSourceFieldArrayFilter, which is also handling field arrays for entity-associated attributes. The reason for this change is when user creates or deletes a group, we shouldn't have to re-build the multiblock model source, all we need is to modify the field data array of the model source.