Removed the non const accessor of AttributeAssignment map.
In Manager and EntityRef classes, there were non-const accessor to the entity-attributes assignment map, which could potenitally cause inconsistency of the map records in the manager and attribute if developers get hold of the map and modify it directly without going through Manager. By removing these non-const accessors, the developers have to go through the public api in manager to update the map so that we can ensure the map is consistent in the system.