Added signal in qtAssociationWidget to broadcast selected entities.
The qtAssociationWidget allows users to select and associate model
entities to attributes. When entities are highlighted in the list
view, it will be nice to show what those entities are in the 3d
render view, hence a signal is added when the entities are selected,
which will allow the application connect and select entity geometry.