Added int property to related models after mesh op
Added an int property, SMTK_MESH_GEN_PROP, to show if the model has
an analysis mesh associated with the model. The value of this property
tells the generation of analysis mesh created for the model. The client
can query this property for the existence of analysis mesh on a model.