Fixed a bug in retranscribing the model after edge split.
After an edge split operation in discrete model, the model was erased from Manager,
then retranscribed into smtk so that the topology and arrangements info are correct.
This should be done efficiently in the future with only local update on related entities.
The bug is that the session was not set for the model after retranscription, which
lead to the session not returning the affected models when asked.