Methods to determine action given click position in model tree
After a mouse click, based on the clicked position, which could be
the eyeball icon for visibility toggle, or the color swatch for
launching the color dialog, the tree view invokes property actions.
In order to record and playback these operations, the recorder
and player need to know the actions being invoked on which index.
This combining with new event translator and player classes makes
recording and playing back changing visibility and colors of
model entities possible.