Modified how entity records are updated for discrete geometry operators
For grow, face merge and split operators in discrete session, to update
all topology and entity relationships properly after entity geometry being
modified or new geometry created, we need to retranscribe the model so
that the records in manager are regenerated. For example, after a face
split or merge, all the related shell use, loop use, and edge use information
need to be updated.