Refactor how discrete session load models and keeps model records
Instead of using a static session object in cmbParserV5, the session object is passed into
the reading operation so that it can be used to assign UUIDs to entities if there are field arrays
for entity UUIDs; if there is no UUID array for certain entities, the session will create
and assign new UUIDs for them. This way cmb models will have consistent UUIDs across
different runs so that entity-attribute associations, which is recorded with UUIDs, will
be consistent, so is color-by entity UUIDs.
NOTE: Currently only vtkCMBParserV5 is using the discrete session